delete function update

+++ mainapp/posts/views.py
@@ -67,6 +67,7 @@ def edit_post(request):
 
 def delete_post(request, post_id=None):
     """configuration of deleting posts page"""
-    post_to_delete = PostForm.objects.get(id=post_id)
-    post_to_delete.delete()
+    if request.method == "DELETE":
+        post_to_delete = PostForm.objects.get(id=post_id)
+        post_to_delete.delete()
     return render(request, 'posts/posts_index.html')
